reminder-service/requirements.txt
supabase==2.6.0
schedule==1.2.0
python-dotenv==1.0.0

